backend/requirements.txt
asgiref==3.9.1
dj-database-url==2.3.0
django==5.0.1
django-tastypie==0.15.0
gunicorn==21.2.0
packaging==23.1
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-mimeparse==1.6.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
whitenoise==6.5.0
django-cors-headers==4.1.0
django-redis==5.4.0
requests==2.32.3
daphne==4.2.1
channels==4.2.2
channels-redis==4.3.0

celery==5.3.6
redis==5.0.1
# django-unfold==0.50.0
django-unfold==0.78.0
django-axes==8.3.1